+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   about schools
 - Added a new list of all projects with Academy URNs
 - Display local authorities and their details
+- Allow local authority records to be edited
 
 ### Changed